+++ packages/riverpod_navigator/example/more_providers.dart
@@ -26,7 +26,7 @@ void main() => runApp(
 // AppNavigator
 //*********************************************
 
-class AppNavigator extends RiverpodRouterDelegate {
+class AppNavigator extends RRouterDelegate {
   AppNavigator(Ref ref, TypedPath homePath) : super(ref, homePath);
 
   @override
@@ -34,13 +34,10 @@ class AppNavigator extends RiverpodRouterDelegate {
     final userIsLogged = ref.read(isLoggedProvider);
 
     // if user is not logged-in and some of the screen in navigations stack needs login => redirect to LoginScreen
-    if (!userIsLogged && ongoingPath.any((segment) => needsLogin(segment)))
-      return [LoginSegment()];
+    if (!userIsLogged && ongoingPath.any((segment) => needsLogin(segment))) return [LoginSegment()];
 
     // user is logged and LogginScreen is going to display => redirect to HomeScreen
-    if (userIsLogged &&
-        (ongoingPath.isEmpty || ongoingPath.last is LoginSegment))
-      return [HomeSegment()];
+    if (userIsLogged && (ongoingPath.isEmpty || ongoingPath.last is LoginSegment)) return [HomeSegment()];
 
     // no redirection is needed
     return ongoingPath;
@@ -48,21 +45,18 @@ class AppNavigator extends RiverpodRouterDelegate {
 }
 
 /// !!! only book screens with odd 'id' require a login
-bool needsLogin(TypedSegment segment) =>
-    segment is BookSegment && segment.id.isOdd;
+bool needsLogin(TypedSegment segment) => segment is BookSegment && segment.id.isOdd;
 
 //*********************************************
 // PROVIDERS
 //*********************************************
 
-final routerDelegateProvider =
-    Provider<AppNavigator>((ref) => AppNavigator(ref, [HomeSegment()]));
+final routerDelegateProvider = Provider<AppNavigator>((ref) => AppNavigator(ref, [HomeSegment()]));
 
 final ongoingPathProvider = StateProvider<TypedPath>((_) => [HomeSegment()]);
 final isLoggedProvider = StateProvider<bool>((_) => false);
 
-final navigationStackProvider =
-    StateProvider<TypedPath>((_) => [HomeSegment()]);
+final navigationStackProvider = StateProvider<TypedPath>((_) => [HomeSegment()]);
 
 //*********************************************
 // MODEL
@@ -73,12 +67,11 @@ typedef JsonMap = Map<String, dynamic>;
 
 /// Common TypedSegment's ancestor
 abstract class TypedSegment {
-  factory TypedSegment.fromJson(JsonMap json) =>
-      json['runtimeType'] == 'BookSegment'
-          ? BookSegment(id: json['id'])
-          : json['runtimeType'] == 'LoginSegment'
-              ? LoginSegment()
-              : HomeSegment();
+  factory TypedSegment.fromJson(JsonMap json) => json['runtimeType'] == 'BookSegment'
+      ? BookSegment(id: json['id'])
+      : json['runtimeType'] == 'LoginSegment'
+          ? LoginSegment()
+          : HomeSegment();
 
   JsonMap toJson() => <String, dynamic>{'runtimeType': runtimeType.toString()};
   @override
@@ -123,7 +116,7 @@ class App extends ConsumerWidget {
 class Defer2NextTick {
   Defer2NextTick();
 
-  late RiverpodRouterDelegate navigator;
+  late RRouterDelegate navigator;
   var _ignoreNextProviderChange = false;
   var _isRunning = false;
 
@@ -133,17 +126,13 @@ class Defer2NextTick {
     _isRunning = true;
     scheduleMicrotask(() {
       try {
-        final navigationStackNotifier =
-            navigator.ref.read(navigationStackProvider.notifier);
-        final ongoingPathNotifier =
-            navigator.ref.read(ongoingPathProvider.notifier);
-        final newNavigStack = navigator.appNavigationLogic(
-            navigationStackNotifier.state, ongoingPathNotifier.state);
+        final navigationStackNotifier = navigator.ref.read(navigationStackProvider.notifier);
+        final ongoingPathNotifier = navigator.ref.read(ongoingPathProvider.notifier);
+        final newNavigStack = navigator.appNavigationLogic(navigationStackNotifier.state, ongoingPathNotifier.state);
         // synchronize navigation stack and ongoingPath
         _ignoreNextProviderChange = true;
         try {
-          ongoingPathNotifier.state =
-              navigationStackNotifier.state = newNavigStack;
+          ongoingPathNotifier.state = navigationStackNotifier.state = newNavigStack;
         } finally {
           _ignoreNextProviderChange = false;
         }
@@ -154,17 +143,13 @@ class Defer2NextTick {
   }
 }
 
-abstract class RiverpodRouterDelegate extends RouterDelegate<TypedPath>
-    with ChangeNotifier, PopNavigatorRouterDelegateMixin<TypedPath> {
-  RiverpodRouterDelegate(this.ref, this.homePath) {
+abstract class RRouterDelegate extends RouterDelegate<TypedPath> with ChangeNotifier, PopNavigatorRouterDelegateMixin<TypedPath> {
+  RRouterDelegate(this.ref, this.homePath) {
     defer2NextTick.navigator = this;
     // listen for "input status" => using defer2NextTick.providerChanged call applicationLogic
-    final unlistens = [ongoingPathProvider, isLoggedProvider]
-        .map((e) => ref.listen(e, (_, __) => defer2NextTick.providerChanged()))
-        .toList();
+    final unlistens = [ongoingPathProvider, isLoggedProvider].map((e) => ref.listen(e, (_, __) => defer2NextTick.providerChanged())).toList();
     // listen navigationStackProvider => call notifyListeners which then rebuilds the navigation stack
-    unlistens
-        .add(ref.listen(navigationStackProvider, (_, __) => notifyListeners()));
+    unlistens.add(ref.listen(navigationStackProvider, (_, __) => notifyListeners()));
     // ignore: avoid_function_literals_in_foreach_calls
     ref.onDispose(() => unlistens.forEach((u) => u()));
   }
@@ -197,19 +182,14 @@ abstract class RiverpodRouterDelegate extends RouterDelegate<TypedPath>
         key: navigatorKey,
         pages: ref
             .read(navigationStackProvider)
-            .map((segment) => MaterialPage(
-                key: ValueKey(segment.toString()),
-                child: screenBuilder(segment)))
+            .map((segment) => MaterialPage(key: ValueKey(segment.toString()), child: screenBuilder(segment)))
             .toList(),
         onPopPage: (route, result) {
           if (!route.didPop(result)) return false;
           final notifier = ref.read(navigationStackProvider.notifier);
           if (notifier.state.length <= 1) return false;
           // remove last segment from navigationStack
-          notifier.state = [
-            for (var i = 0; i < notifier.state.length - 1; i++)
-              notifier.state[i]
-          ];
+          notifier.state = [for (var i = 0; i < notifier.state.length - 1; i++) notifier.state[i]];
           return true;
         });
   }
@@ -221,15 +201,11 @@ abstract class RiverpodRouterDelegate extends RouterDelegate<TypedPath>
     return SynchronousFuture(null);
   }
 
-  void navigate(TypedPath newPath) =>
-      ref.read(ongoingPathProvider.notifier).state = newPath;
+  void navigate(TypedPath newPath) => ref.read(ongoingPathProvider.notifier).state = newPath;
 
   void replaceLast<T extends TypedSegment>(T Function(T old) replace) {
     final navigationStack = ref.read(navigationStackProvider);
-    return navigate([
-      for (var i = 0; i < navigationStack.length - 1; i++) navigationStack[i],
-      replace(navigationStack.last as T)
-    ]);
+    return navigate([for (var i = 0; i < navigationStack.length - 1; i++) navigationStack[i], replace(navigationStack.last as T)]);
   }
 }
 
@@ -239,19 +215,14 @@ abstract class RiverpodRouterDelegate extends RouterDelegate<TypedPath>
 
 class RouteInformationParserImpl implements RouteInformationParser<TypedPath> {
   @override
-  Future<TypedPath> parseRouteInformation(RouteInformation routeInformation) =>
-      Future.value(path2TypedPath(routeInformation.location));
+  Future<TypedPath> parseRouteInformation(RouteInformation routeInformation) => Future.value(path2TypedPath(routeInformation.location));
 
   @override
-  RouteInformation restoreRouteInformation(TypedPath configuration) =>
-      RouteInformation(location: typedPath2Path(configuration));
+  RouteInformation restoreRouteInformation(TypedPath configuration) => RouteInformation(location: typedPath2Path(configuration));
 
-  static String typedPath2Path(TypedPath typedPath) => typedPath
-      .map((s) => Uri.encodeComponent(jsonEncode(s.toJson())))
-      .join('/');
+  static String typedPath2Path(TypedPath typedPath) => typedPath.map((s) => Uri.encodeComponent(jsonEncode(s.toJson()))).join('/');
 
-  static String debugTypedPath2Path(TypedPath typedPath) =>
-      typedPath.map((s) => jsonEncode(s.toJson())).join('/');
+  static String debugTypedPath2Path(TypedPath typedPath) => typedPath.map((s) => jsonEncode(s.toJson())).join('/');
 
   static TypedPath path2TypedPath(String? path) {
     if (path == null || path.isEmpty) return [];
@@ -308,8 +279,7 @@ class HomeScreen extends ConsumerWidget {
                       HomeSegment(),
                       BookSegment(id: i),
                     ]),
-                    child: Text(
-                        'Go to Book $i ${i.isOdd && !isLogged ? ' (needs login)' : ''}'),
+                    child: Text('Go to Book $i ${i.isOdd && !isLogged ? ' (needs login)' : ''}'),
                   );
                 }),
               ]
@@ -336,16 +306,12 @@ class BookScreen extends ConsumerWidget {
             children: [
               const SizedBox(height: 30),
               ElevatedButton(
-                onPressed: () => ref
-                    .read(routerDelegateProvider)
-                    .replaceLast<BookSegment>(
-                        (old) => BookSegment(id: old.id + 1)),
+                onPressed: () => ref.read(routerDelegateProvider).replaceLast<BookSegment>((old) => BookSegment(id: old.id + 1)),
                 child: const Text('Go to next book'),
               ),
               const SizedBox(height: 30),
               ElevatedButton(
-                onPressed: () =>
-                    ref.read(routerDelegateProvider).navigate([HomeSegment()]),
+                onPressed: () => ref.read(routerDelegateProvider).navigate([HomeSegment()]),
                 child: const Text('Go to home'),
               ),
             ],
@@ -364,8 +330,7 @@ class LoginScreen extends ConsumerWidget {
         appBar: AppBar(
           title: const Text('Login Screen'),
           leading: IconButton(
-            onPressed: () =>
-                ref.read(routerDelegateProvider).navigate([HomeSegment()]),
+            onPressed: () => ref.read(routerDelegateProvider).navigate([HomeSegment()]),
             icon: const Icon(Icons.cancel),
           ),
         ),
@@ -376,9 +341,7 @@ class LoginScreen extends ConsumerWidget {
               const SizedBox(height: 30),
               ElevatedButton(
                 onPressed: () {
-                  ref.read(ongoingPathProvider.notifier).state = [
-                    HomeSegment()
-                  ];
+                  ref.read(ongoingPathProvider.notifier).state = [HomeSegment()];
                   ref.read(isLoggedProvider.notifier).state = true;
                 },
                 child: const Text('Login'),